requirements.txt
chromadb
faiss-cpu
google-generativeai
langchain
langchain_community
langchain_google_genai
PyPDF2
python-dotenv
python-pptx
streamlit